requirements.txt
Django==1.11
psycopg2>=2.6.2
djangorestframework>=3.5.3
dj-database-url
gunicorn
coreapi
django-dash
django-webpack-loader
python-dateutil
